requirements.txt
# --- Core Frameworks ---
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0

# --- IA & Data (RAG) ---
langchain>=0.0.339
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.0.1
ollama>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# --- Base de Datos (PostgreSQL) ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# --- Utilidades ---
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
bcrypt>=4.0.1
bcrypt>=4.0.0
